requirements.txt
docker
molecule
molecule-docker
pytest
pytest-pretty
pytest-testinfra
pyyaml
